move filterWIndex to spark3.0+ because previous versions are unavailable
