+++ core/src/test/spark_3.0_3.1_3.2_3.3/scala/doric/syntax/ArrayColumns3xSpec.scala
@@ -1,7 +1,8 @@
 package doric
 package syntax
 
-import org.apache.spark.sql.catalyst.expressions.ArraySort
+import doric.SparkAuxFunctions.createLambda
+import org.apache.spark.sql.catalyst.expressions.{ArrayFilter, ArraySort}
 import org.scalatest.EitherValues
 import org.scalatest.matchers.should.Matchers
 import org.apache.spark.sql.{Column, functions => f}
@@ -22,6 +23,40 @@ class ArrayColumns3xSpec
         c => colArrayString(c).forAll(_.isNotNull),
         c => f.forall(f.col(c), _.isNotNull),
         List(Some(false), Some(true), None)
+      )
+    }
+  }
+
+  describe("filterWIndex doric function") {
+    import spark.implicits._
+
+    it("should work as spark filter((Column, Column) => Column) function") {
+      val df = List((Array("a", "b", "c", "d"), "b"))
+        .toDF("col1", "col2")
+
+      noException shouldBe thrownBy {
+        df.select(colArrayString("col1").filterWIndex((x, i) => {
+          i === 0.lit or x === colString("col2")
+        }))
+      }
+
+      lazy val filter_old: (Column, (Column, Column) => Column) => Column =
+        (myCol, myFun) =>
+          new Column(ArrayFilter(myCol.expr, createLambda(myFun)))
+
+      df.testColumns2("col1", "col2")(
+        (c1, c2) =>
+          colArrayString(c1).filterWIndex((x, i) => {
+            i === 0.lit or x === colString(c2)
+          }),
+        (c1, c2) =>
+          filter_old(
+            f.col(c1),
+            (x, i) => {
+              i === 0 or x === f.col(c2)
+            }
+          ),
+        List(Some(Array("a", "b")))
       )
     }
   }
